rag_service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai==1.58.1
faiss-cpu==1.9.0.post1
numpy==2.2.1
python-dotenv==1.0.1
